feat: [023-REFACTOR-CUSTOMER-SERVICE]
store-service 리펙토링
