--- domain/src/main/java/com/project/domain/DomainApplication.java
@@ -1,12 +1,0 @@
-package com.project.domain;
-
-import org.springframework.boot.SpringApplication;
-
-//@SpringBootApplication
-public class DomainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DomainApplication.class, args);
-    }
-
-}
