--- store-service/src/main/java/com/project/storeservice/StoreServiceApplication.java
@@ -1,13 +1,0 @@
-package com.project.storeservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class StoreServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(StoreServiceApplication.class, args);
-    }
-
-}
